feat: add option to control whether console visible or not; add build configure to support no-consle mode; type-safe problem consideratino for 64-bit
more:
1. refactor the mechanism of handling exceptions
2. split code of reading metadata
3. optimize code format

+++ Project1/debug.cpp
@@ -22,7 +22,7 @@ string get_last_error_message()
 	string result;
 
 	char* msg_buf;
-	uint64_t err_code = GetLastError();
+	DWORD err_code = GetLastError();
 
 	FormatMessageA(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM,
 		0,
